+++ Runtime/Extensions/Color/ColorComparisonExtensions.cs
@@ -13,6 +13,14 @@ namespace LiteNinja.Colors.Extensions
                    Mathf.Approximately(value.b, other.b) &&
                    Mathf.Approximately(value.a, other.a);
         }
+        
+
+        public static bool ApproximatelyRGB(this Color value, Color other)
+        {
+            return Mathf.Approximately(value.r, other.r) &&
+                   Mathf.Approximately(value.g, other.g) &&
+                   Mathf.Approximately(value.b, other.b);
+        }
 
         public static bool Approximately(this Color value, Color other, float epsilon)
         {
@@ -20,6 +28,13 @@ namespace LiteNinja.Colors.Extensions
                    value.g.Approximately(other.g, epsilon) &&
                    value.b.Approximately(other.b, epsilon) &&
                    value.a.Approximately(other.a, epsilon);
+        }
+        
+        public static bool ApproximatelyRGB(this Color value, Color other, float epsilon)
+        {
+            return value.r.Approximately(other.r, epsilon) &&
+                   value.g.Approximately(other.g, epsilon) &&
+                   value.b.Approximately(other.b, epsilon);
         }
 
         public static bool IsApproximatelyBlack(this Color self)