+++ Runtime/Extensions/Color/ColorGrayscaleExtensions.cs
@@ -15,7 +15,7 @@ namespace LiteNinja.Colors.Extensions
         /// <summary>
         /// Returns a new instance of the color with zero saturation.
         /// </summary>
-        public static Color Grayscaled(this Color color)
+        public static Color Grayscale(this Color color)
         {
             return color.Desaturate(1f);
         }