+++ Runtime/Extensions/Color/ColorHarmoniesExtensions.cs
@@ -1,0 +1,18 @@
+using System.Collections.Generic;
+using LiteNinja.Colors.Spaces;
+using UnityEngine;
+
+namespace LiteNinja.Colors.Extensions
+{
+    public static class ColorHarmoniesExtensions
+    {
+        /// <summary>
+        /// Creates a monochromatic harmony of the color, a set of tints and shades formed from the base color.
+        /// </summary>
+        public static IEnumerable<Color> MonochromaticHarmony(this Color color, int count)
+        {
+
+            return null;
+        }
+    }
+}