+++ CHANGELOG.md
@@ -8,9 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Color Extensions for Contrast and Hue
 - Color Extensions for Saturation and Lightness
 - Color Extensions for Color blindness
-
+- Color Extensions for Gradients
+- Color Extensions for reducing number of colors to a minimun number or merging together similar colors
+- Color Extensions for Lerp (with Tint, Shade and Tones)
 ### Changed
 - Split ColorExtensions in multiple files by common theme
+- IPalette definition
+
+### Removed
+- Removed Palette Generators for simpler extension methods
 
 ## [1.1.2] - 2022-07-07
 ### Added